2022_10_3_Zapata_Toffoli_test_circuits/requirements.txt
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cirq-core==1.0.0
cirq-google==1.0.0
cmake==3.24.1.1
contourpy==1.0.5
cryptography==38.0.1
cycler==0.11.0
deprecation==2.1.0
dill==0.3.5.1
duet==0.2.7
fonttools==4.37.4
gcc7==0.0.7
google-api-core==1.33.1
google-auth==2.12.0
googleapis-common-protos==1.56.4
grpcio==1.49.1
grpcio-status==1.48.2
h5py==3.7.0
idna==3.4
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.6.0
mpmath==1.2.1
networkx==2.8.7
ntlm-auth==1.5.0
numpy==1.23.3
openfermion==1.5.1
orquestra-cirq==0.5.0
orquestra-core==0.5.0
orquestra-opt==0.5.0
orquestra-qiskit==0.5.0
orquestra-quantum==0.5.0
orquestra-qulacs==0.4.0
orquestra-vqa==0.5.0
overrides==6.2.0
packaging==21.3
pandas==1.5.0
pbr==5.10.0
Pillow==9.2.0
ply==3.11
proto-plus==1.22.1
protobuf==3.20.3
psutil==5.9.2
PubChemPy==1.0.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyparsing==3.0.9
python-constraint==1.4.0
python-dateutil==2.8.2
python-rapidjson==1.8
pytz==2022.4
qiskit==0.34.0
qiskit-aer==0.10.1
qiskit-ibmq-provider==0.18.3
qiskit-ignis==0.7.0
qiskit-terra==0.19.1
Qulacs==0.2.0
requests==2.28.1
requests-ntlm==1.1.0
retworkx==0.11.0
rsa==4.9
scikit-learn==1.1.2
scipy==1.9.1
six==1.16.0
sortedcontainers==2.4.0
stevedore==4.0.0
symengine==0.9.2
sympy==1.9
threadpoolctl==3.1.0
tqdm==4.64.1
tweedledum==1.1.1
typing_extensions==4.3.0
urllib3==1.26.12
websocket-client==1.4.1
